vertex_ai/requirements.txt
# Vertex AI and Google Cloud
google-cloud-aiplatform>=1.36.0
google-cloud-storage>=2.10.0

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
catboost>=1.2.0
shap>=0.42.0

# NLP for skill similarity
sentence-transformers>=2.2.0

# Utilities
python-multipart
pydantic>=2.0.0
python-dotenv